+++ src/index.ts
@@ -1,7 +1,7 @@
 /*! koilib - MIT License (c) Julian Gonzalez (joticajulian@gmail.com) */
 // eslint-disable import/no-cycle
 export * as utils from "./indexUtils";
-export * as interfaces from "./interface";
+export * from "./interface";
 export * from "./Contract";
 export * from "./Signer";
 export * from "./Provider";